networksecurity/components/data_validation.py
# We are going to read the path given by data ingestion artifact
from networksecurity.entity.artifact_entity import DataIngestionArtifact,DataValidationArtifact
from networksecurity.entity.config_entity import DataValidationConfig
from networksecurity.exception.exception import NetworkSecurtiyException
from networksecurity.logging.logger import logging
from networksecurity.constant.training_pipeline import SCHEMA_FILE_PATH
from networksecurity.utils.main_utils.utils import read_yaml_file

from scipy.stats import ks_2samp
# ks_2samp find out weather there is data drift or not
import pandas as pd
import os,sys

# Reading the Schema and make comparitions
# If the comparison is successfull then it will say that training status is true
# columns are not  missing in the dataframe 

class DataValidation:
    def __init__(self,data_ingestion_artifact:DataIngestionArtifact,
                 data_validation_config:DataValidationConfig):
        try:
            self.data_ingestion_artifact=data_ingestion_artifact
            self.data_validation_config=data_validation_config
            self._schema_config=read_yaml_file(SCHEMA_FILE_PATH)
        except Exception as e:
            raise NetworkSecurtiyException(e,sys)